+++ markup/list-loading.html
@@ -29,7 +29,7 @@
               <a class="trip-tabs__btn" href="#">Stats</a>
             </nav>
 
-            <h2 class="visually-hidden">Filter events</h2>
+            <h2 class="visually-hidden">FilterTpl events</h2>
             <form class="trip-filters" action="#" method="get">
               <div class="trip-filters__filter">
                 <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
