--- src/components/cost-info.js
@@ -1,7 +1,0 @@
-export const createCostInfoTemplate = () => {
-  return (
-    ` <p class="trip-info__cost">
-      Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
-    </p>`
-  );
-};
