+++ src/components/cost/cost-tpl.js
@@ -1,0 +1,7 @@
+const createCostTemplate = () => {
+  return ` <p class="trip-info__cost">
+            Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
+          </p>`;
+};
+
+export {createCostTemplate};
