+++ src/components/filters/filters.js
@@ -1,0 +1,25 @@
+import {createElement} from "../../utils";
+import {createFiltersTemplate} from "./filters-tpl.js";
+
+export default class FilterTpl {
+  constructor(filters) {
+    this._filters = filters;
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createFiltersTemplate(this._filters);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
