+++ src/components/menu/menu.js
@@ -1,0 +1,24 @@
+import {createElement} from "../../utils";
+import {createMenuTemplate} from "./menu-tpl.js";
+
+export default class Menu {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createMenuTemplate();
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
