+++ src/components/sort/sort.js
@@ -1,0 +1,24 @@
+import {createElement} from "../../utils";
+import {createSortTemplate} from "./sort-tpl";
+
+export default class Sort {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createSortTemplate();
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
