+++ src/components/trip-days/trip-days-tpl.js
@@ -1,0 +1,5 @@
+const createTripDaysTemplate = () => {
+  return `<ul class="trip-days"></ul>`;
+};
+
+export {createTripDaysTemplate};
