--- src/components/trip-events-list.js
@@ -1,3 +1,0 @@
-export const createTripEventsListTemplate = () => {
-  return `<ul class="trip-events__list"></ul>`;
-};
