+++ src/components/trip-events-list/trip-events-list-tpl.js
@@ -1,0 +1,5 @@
+const createTripEventsListTemplate = () => {
+  return `<ul class="trip-events__list"></ul>`;
+};
+
+export {createTripEventsListTemplate};
