+++ src/components/trip-events/trip-events-item-tpl.js
@@ -1,0 +1,5 @@
+const createTripEventsItemTemplate = () => {
+  return `<li class="trip-events__item"></li>`;
+};
+
+export {createTripEventsItemTemplate};
