+++ src/components/trip-info/trip-info.js
@@ -1,0 +1,24 @@
+import {createElement} from "../../utils";
+import {createTripInfoTemplate} from "./trip-info-tpl";
+
+export default class TripInfo {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createTripInfoTemplate();
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
